New tool - rfid.sh
This tool will handle the RFID database, deprecating the MySQL version in io
folder (blackknightdb).

The backend is LDAP based. This means every HSBXL user will be able to manage
(add, view, delete) their own RFID credentials.

For the tag to be active and valid, the user must be part of both the member
group and the access control group.

At the moment, the tool can only dump the tags from the database, for use with
an access controller. Tag handling is not implemented yet.